geofr/requirements.txt
torch==2.6.0
typer==0.15.2
fastapi==0.115.12
uvicorn==0.34.2
matplotlib
numpy
scikit-learn
wandb
pandas>=2.2.0
seaborn>=0.13.0
imbalanced-learn>=0.12.0
pytest==8.3.4
invoke==2.2.0
pytest==8.3.4
ruff==0.1.3
nltk>=3.9.1
gensim>=4.3.0
pyLDAvis>=3.4.1
kaggle
hydra-core>=1.3.2
omegaconf>=2.3.0
loguru
wandb
python-dotenv
